+++ custom_components/homematicip_local/climate.py
@@ -323,16 +323,30 @@ class AioHomematicClimate(AioHomematicGenericRestoreEntity[BaseCustomDpClimate],
     @handle_homematic_errors
     async def async_get_schedule_profile(self, profile: str | ScheduleProfile) -> ServiceResponse:
         """Get a schedule profile."""
+        _LOGGER.warning("Service 'get_schedule_profile' is deprecated and will be removed in April 2026. ")
         schedule_profile = profile if isinstance(profile, ScheduleProfile) else ScheduleProfile(profile)
         if profile_data := await self._data_point.get_schedule_profile(profile=schedule_profile):
             return cast(ServiceResponse, profile_data)
         return None
 
     @handle_homematic_errors
+    async def async_get_schedule_simple_profile(self, profile: ScheduleProfile) -> ServiceResponse:
+        """Get a schedule simple profile."""
+        return cast(ServiceResponse, await self._data_point.get_schedule_simple_profile(profile=profile))
+
+    @handle_homematic_errors
+    async def async_get_schedule_simple_weekday(self, profile: ScheduleProfile, weekday: WeekdayStr) -> ServiceResponse:
+        """Get a schedule simple profile weekday."""
+        return cast(
+            ServiceResponse, await self._data_point.get_schedule_simple_weekday(profile=profile, weekday=weekday)
+        )
+
+    @handle_homematic_errors
     async def async_get_schedule_weekday(
         self, profile: str | ScheduleProfile, weekday: str | WeekdayStr
     ) -> ServiceResponse:
         """Get a schedule profile weekday."""
+        _LOGGER.warning("Service 'get_schedule_weekday' is deprecated and will be removed in April 2026.")
         schedule_profile = profile if isinstance(profile, ScheduleProfile) else ScheduleProfile(profile)
         schedule_weekday = weekday if isinstance(weekday, WeekdayStr) else WeekdayStr(weekday)
         if weekday_data := await self._data_point.get_schedule_weekday(